testing file input

+++ project/seats/flightseats/test_flightseats/test_admin.py
@@ -1,3 +1,4 @@
+from django import forms
 from django.test import RequestFactory, TestCase
 from django.urls import reverse
 from ..admin import SeatAdmin
@@ -7,14 +8,11 @@ import os
 class SeatAdminTestCase(TestCase):
     def test_upload_txt(self):
         # create a fake request object with a POST request
-        factory = RequestFactory()
-        txt_file = open('file.txt', 'w')
-        txt_file.write('1    A    B    C    D    E    F\n'
-                       '2    A    B    C    D    E    F\n'
-                       '3    A    B    C    D    E    F')
+        txt_data = '1\tA\tB\tC\tD\tE\tF\n' \
+                   '2\tA\tB\tC\tD\tE\tF\n' \
+                   '3\tA\tB\tC\tD\tE\tF'
 
-        request = factory.post('/admin/flightseats/book/txt-upload/', {'txt_upload': txt_file})
-        txt_file.close()
+        request = self._create_post_request(txt_data)
 
         # create an instance of the SeatAdmin class
         admin = SeatAdmin()
@@ -26,11 +24,20 @@ class SeatAdminTestCase(TestCase):
         self.assertEqual(response.status_code, 200)
 
         # check that the file has been written with the correct data
-        with open('flightseats/data/chartIn.txt', 'r') as f:
-            content = f.read()
-            self.assertEqual(content, '1    A    B    C    D    E    F\n'
-                                      '2    A    B    C    D    E    F\n'
-                                      '3    A    B    C    D    E    F')
+        content = self._read_file_content()
+        self.assertEqual(content, txt_data)
 
+    def _create_post_request(self, txt_data):
+        factory = RequestFactory()
+        txt_file = SimpleUploadedFile("file.txt", txt_data.encode())
+        return factory.post('/admin/flightseats/book/txt-upload/', {'txt_upload': txt_file})
+
+    def _read_file_content(self):
+        with open('flightseats/data/chartIn.txt', 'r') as f:
+            return f.read()
+
+    def tearDown(self):
         # clean up the file after the test
-        os.remove('flightseats/data/chartIn.txt')
+        if os.path.exists('flightseats/data/chartIn.txt'):
+            os.remove('flightseats/data/chartIn.txt')
+
